Return java byte with byteValue method
Ballerina byte representation will be from 0-255, but the bit
pattern will be same in both java and ballerina. So returning
java byte with byteValue method and use Byte.unsignedInt to
represent ballerina byte.

+++ bvm/ballerina-core/src/main/java/org/ballerinalang/model/values/BBlob.java
@@ -55,7 +55,7 @@ public class BBlob extends BValueType implements BRefType<byte[]> {
     }
 
     @Override
-    public int byteValue() {
+    public byte byteValue() {
         return 0;
     }
