+++ bvm/ballerina-core/src/main/java/org/ballerinalang/model/values/BBoolean.java
@@ -51,7 +51,7 @@ public final class BBoolean extends BValueType implements BRefType<Boolean> {
     }
 
     @Override
-    public int byteValue() {
+    public byte byteValue() {
         return 0;
     }
