+++ bvm/ballerina-core/src/main/java/org/ballerinalang/util/codegen/cpentries/ByteCPEntry.java
@@ -55,6 +55,6 @@ public class ByteCPEntry implements ConstantPoolEntry {
 
     @Override
     public int hashCode() {
-        return Integer.hashCode(value);
+        return Byte.hashCode(value);
     }
 }
